Add upload profile and changed book image

+++ src/main/java/com/libraryrest/DAO/UserDao.java
@@ -9,7 +9,7 @@ public interface UserDao {
 
     public List<User> findAll();
 
-    public User findById(Long id);
+    public User findById(Integer id);
 
     public User findByName(String login);
 
@@ -21,5 +21,5 @@ public interface UserDao {
     public void remove(User user);
 
     @Secured({"ROLE_ADMIN"})
-    public void remove(Long id);
+    public void remove(Integer id);
 }
