+++ src/main/java/com/libraryrest/DAOImpl/UserDaoImpl.java
@@ -29,7 +29,7 @@ public class UserDaoImpl implements UserDao {
 
     @Override
     @Transactional
-    public User findById(Long userId) {
+    public User findById(Integer userId) {
         String hql = "from User u where u.id = :userId";
         Query query = sessionFactory.getCurrentSession().createQuery(hql);
         query.setParameter("userId", userId);
@@ -68,7 +68,7 @@ public class UserDaoImpl implements UserDao {
 
     @Override
     @Transactional
-    public void remove(Long id) {
+    public void remove(Integer id) {
         User user = findById(id);
         sessionFactory.getCurrentSession().delete(user);
     }
